feat: check_duplication 함수 작성#17
NULL과 -1로 초기화한 뒤 중복 되는 identifier를 만났을 시 실패 분기로
빠지는 함수 작성 후, normenette를 맞췄다.
