+++ srcs/parse_map/example.cub
@@ -1,13 +1,15 @@
 C 224,2,2
 
 
+
 SO ./path_to_the_south_texture
+
+NO           ./path_to_the_west_texture
+
+
 WE           ./path_to_the_west_texture
 
 F 220,100,0
-
-WE           ./path_to_the_west_texture
-
 
 
 EA ./path_to_the_east_texture
