+++ srcs/parse_map/init_map_info.c
@@ -6,7 +6,7 @@
 /*   By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/17 13:04:54 by jayoon            #+#    #+#             */
-/*   Updated: 2022/10/22 22:14:04 by jayoon           ###   ########.fr       */
+/*   Updated: 2022/10/24 13:24:49 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,205 +33,6 @@ static int	safe_open(char *file_path)
 	return (fd);
 }
 
-static void	init_texture_path_variable(t_substr_info *info, char **p_str)
-{
-	info->start = NULL;
-	info->end = NULL;
-	info->flag = 0;
-	*p_str = *p_str + 3;
-}
-
-static t_identifier	process_texture_path(t_map_info *map_info, char *str, \
-						t_identifier num_iden)
-{
-	t_substr_info	info;
-
-	init_texture_path_variable(&info, &str);
-	while (*str)
-	{
-		if (*str != ' ')
-		{
-			if (info.end != NULL)
-				print_error_str("Too many information!\n");
-			if (info.flag == 0)
-			{
-				info.start = str;
-				info.flag = 1;
-			}
-			if (*(str + 1) == '\0' || *(str + 1) == ' ')
-				info.end = str;
-		}
-		++str;
-	}
-	if (info.start != NULL)
-		map_info->texture_path[num_iden] \
-			= ft_substr(info.start, 0, (info.end - info.start) + 1);
-	else
-		print_error_str("There is not information!\n");
-	return (num_iden);
-}
-
-static void	init_color_variable(t_substr_info *info, char **p_str)
-{
-	info->start = NULL;
-	info->end = NULL;
-	info->flag = 0;
-	*p_str = *p_str + 2;
-}
-
-static char	*pass_space(char *str)
-{
-	while (*str == ' ')
-		++str;
-	if (*str == '\0')
-		return (NULL);
-	return (str);
-}
-
-static size_t	count_comma(char *str)
-{
-	size_t	cnt;
-
-	cnt = 0;
-	while (*str)
-	{
-		if (*str == ',')
-			++cnt;
-		++str;
-	}
-	return (cnt);
-}
-
-static void	check_0_to_255(int num)
-{
-	if (num < 0 || num > 255)
-		print_error_str("Invalid nubmer!\n");
-}
-
-static void	get_color(t_map_info *map_info, t_identifier num_iden, char *str)
-{
-	char			**arr_rgb;
-	t_color			color;
-	int				num;
-
-	color = RED;
-	arr_rgb = ft_split(str, ',');
-	while (*arr_rgb)
-	{
-		num = cub3d_atoi(*arr_rgb);
-		if (num == -1)
-			print_error_str("Invalid number!\n");
-		check_0_to_255(num);
-		if (num_iden == FLOOR)
-			map_info->floor[color] = num;
-		else
-			map_info->ceiling[color] = num;
-		++color;
-		++arr_rgb;
-	}
-	if (color != 3)
-		print_error_str("Invalid information!\n");
-}
-
-/*
- * Format is [  "F or C"  R,G,B  ]
- */
-static t_identifier	process_color(t_map_info *map_info, char *str, \
-						t_identifier num_iden)
-{
-	t_substr_info	info;
-	size_t			cnt_comma;
-
-	init_color_variable(&info, &str);
-	str = pass_space(str);
-	if (str == NULL)
-		return (ELEMENT_FAIL);
-	cnt_comma = count_comma(str);
-	if (cnt_comma == 2)
-		get_color(map_info, num_iden, str);
-	else
-		print_error_str("Invalid information!\n");
-	return (num_iden);
-}
-
-static t_identifier	check_identifier(t_map_info *map_info, char *str)
-{
-	const t_identifier_info	arr[6] = {{"EA ", 3, EAST, process_texture_path}, \
-									{"WE ", 3, WEST, process_texture_path}, \
-									{"SO ", 3, SOUTH, process_texture_path}, \
-									{"NO ", 3, NORTH, process_texture_path}, \
-									{"F ", 2, FLOOR, process_color}, \
-									{"C ", 2, CEILING, process_color}};
-	int						i;
-
-	i = 0;
-	while (i < 6)
-	{
-		if (!ft_strncmp(str, arr[i].str, arr[i].len))
-			return (arr[i].f_parsing(map_info, str, arr[i].identifier));
-		++i;
-	}
-	return (ELEMENT_FAIL);
-}
-
-static void	change_eol_to_nul(char *str)
-{
-	while (*str)
-	{
-		if (*str == '\n')
-			*str = '\0';
-		++str;
-	}
-}
-
-/*
- * First word followd zero or some space is identifier.
- */
-static t_identifier 	parse_string(t_map_info *map_info, char *str)
-{
-	t_identifier	identifier;
-
-	identifier = ELEMENT_FAIL;
-	change_eol_to_nul(str);
-	while (*str)
-	{
-		if (*str != ' ')
-			return (check_identifier(map_info, str));
-		++str;
-	}
-	return (identifier);
-}
-
-/*
- * If first character is not eol, that identifier will be checked.
- */
-static void	init_element(t_map_info *map_info, int fd)
-{
-	char	*str;
-	size_t	cnt_element;	
-
-	cnt_element = 0;
-	while (1)
-	{
-		str = get_next_line(fd);
-		if (str == NULL)
-			break ;
-		if (*str != '\n')
-		{
-			if (parse_string(map_info, str) == ELEMENT_FAIL)
-				print_error_str("Invalid identifier\n");
-			++cnt_element;
-		}
-		free(str);
-		str = NULL;
-		if (cnt_element > 6)
-			break ;
-	}
-	// 중복 없이 모든 식별자를 읽었는지 확인한다.
-	// 그렇다면 map_info의 path가 NULL이 아니다.
-	// ceiling과 floor의 모든 값이 -1이 아니다.
-}
-
 static void	init_map_info_to_initial_value(t_map_info *map_info)
 {
 	size_t	i;
@@ -249,6 +50,28 @@ static void	init_map_info_to_initial_value(t_map_info *map_info)
 	}
 }
 
+static void	check_duplication(t_map_info *map_info)
+{
+	size_t	i;
+	int		flag_fail;
+
+	i = 0;
+	flag_fail = 0;
+	while (i < 4)
+	{
+		if (map_info->texture_path[i] == NULL)
+			flag_fail = 1;
+		if (i < 3)
+		{
+			if (map_info->ceiling[i] == -1 || map_info->floor[i] == -1)
+				flag_fail = 1;
+		}
+		++i;
+	}
+	if (flag_fail == 1)
+		print_error_str("Identifier is duplication!\n");
+}
+
 void	init_map_info(t_map_info *map_info, int argc, char *file_path)
 {
 	int			fd;
@@ -258,6 +81,11 @@ void	init_map_info(t_map_info *map_info, int argc, char *file_path)
 	fd = safe_open(file_path);
 	init_map_info_to_initial_value(map_info);
 	init_element(map_info, fd);
+	check_duplication(map_info);
 	// init_map_content(map_info, fd);
 	close(fd);
 }
+
+	// 중복 없이 모든 식별자를 읽었는지 확인한다.
+	// 그렇다면 map_info의 path가 NULL이 아니다.
+	// ceiling과 floor의 모든 값이 -1이 아니다.
